[feat]: LadderController 생성
